+++ tests/manual/rhi/msaatexture/msaatexture.cpp
@@ -171,7 +171,7 @@ void Window::customInit()
     inputLayout.setBindings({ { 4 * sizeof(float) } });
     inputLayout.setAttributes({
         { 0, 0, QRhiVertexInputAttribute::Float2, 0 },
-        { 0, 1, QRhiVertexInputAttribute::Float2, 2 * sizeof(float) }
+        { 0, 1, QRhiVertexInputAttribute::Float2, quint32(2 * sizeof(float)) }
     });
     d.psLeft->setVertexInputLayout(inputLayout);
     d.psLeft->setShaderResourceBindings(d.srbLeft);
@@ -227,7 +227,7 @@ void Window::customInit()
     });
     inputLayout.setAttributes({
         { 0, 0, QRhiVertexInputAttribute::Float2, 0 },
-        { 0, 1, QRhiVertexInputAttribute::Float3, 2 * sizeof(float) }
+        { 0, 1, QRhiVertexInputAttribute::Float3, quint32(2 * sizeof(float)) }
     });
     d.triPs->setVertexInputLayout(inputLayout);
     d.triPs->setShaderResourceBindings(d.triSrb);
@@ -299,7 +299,7 @@ void Window::customRender()
     cb->setGraphicsPipeline(d.triPs);
     cb->setViewport({ 0, 0, float(d.msaaTex->pixelSize().width()), float(d.msaaTex->pixelSize().height()) });
     cb->setShaderResources();
-    QRhiCommandBuffer::VertexInput vbufBinding(d.vbuf, sizeof(vertexData));
+    QRhiCommandBuffer::VertexInput vbufBinding(d.vbuf, quint32(sizeof(vertexData)));
     cb->setVertexInput(0, 1, &vbufBinding);
     cb->draw(3);
     cb->endPass();
